Move `FromStr` impl to `datetime` module
